[addync_sync_to_suggestions] Adding sync to Ruby suggestions
