+++ src/generator/suggestions/ruby21.js
@@ -1,8 +1,7 @@
 import { _ } from 'azk';
-import { UIProxy } from 'azk/cli/ui';
-import { example_system } from 'azk/generator/rules';
+import { Suggestion as DefaultSuggestion } from 'azk/generator/suggestions/ruby_default';
 
-export class Suggestion extends UIProxy {
+export class Suggestion extends DefaultSuggestion {
   constructor(...args) {
     super(...args);
 
@@ -13,23 +12,9 @@ export class Suggestion extends UIProxy {
     this.ruleNamesList = ['ruby21'];
 
     // Initial Azkfile.js suggestion
-    this.suggestion = _.extend({}, example_system, {
+    this.suggestion = _.extend({}, this.suggestion, {
       __type  : 'ruby 2.1',
       image   : { docker: 'azukiapp/ruby:2.1' },
-      provision: [
-        'bundle install --path /azk/bundler',
-      ],
-      http    : true,
-      scalable: { default: 2 },
-      command : 'bundle exec rackup config.ru --pid /tmp/ruby.pid --port $HTTP_PORT --host 0.0.0.0',
-      mounts  : {
-        '/azk/#{manifest.dir}': {type: 'path', value: '.'},
-        '/azk/bundler'        : {type: 'persistent', value: 'bundler'},
-      },
-      envs    : {
-        RUBY_ENV : 'development',
-        BUNDLE_APP_CONFIG : '/azk/bundler',
-      }
     });
   }
